refactor: split center command family
